Extract clustering assignments.

This function extracts the clustering assignments and the latent variables from the fit. This function must receive in input exactly the same data used for the fit.

Clusters(x, data)

Arguments

x
An object of class bmix that represents a fit.

data
The data used to compute the fit x.

Value

The input data transformed into a tibble, augmented with a column for the cluster labels and the latent variables.

Examples

# The same dataset used in the package vignette
data = data.frame(successes = c(rbinom(30, 100, .4), rbinom(70, 100, .7)), trials = 100)

# BMix fit with default parameters
x = bmixfit(data)

#>  [ BMix ~ Fitting data ] 
#> Total number of runs:  4 
#> 
#>      Run #  | Samp. # |  Binom. | Beta-Binom. |      Conv. |        ICL
#> ------------------------------------------------------------------------------
#>        1/4  |       1 |       1 |           0 |         OK | 1447.59664233386 *
#>        2/4  |       2 |       1 |           0 |         OK | 1427.338531978 *
#>        3/4  |       1 |       2 |           0 |         OK | 730.335117206374 *
#>        4/4  |       2 |       2 |           0 |         OK | 749.7597
#> 
#>  *** Best with ICL:  730.3351  
#> 
#> BMix model with K = 2 components: 2 Binomials and 0 Beta-Binomials.
#> 
#> Binomials
#>     Bin 1     Bin 2 
#> 0.7058524 0.3979979 
#> 
#> Beta-Binomials
#> NULL
#> 
#> ICL: 730.3351 

Clusters(x, data)

#> # A tibble: 100 x 5
#>    successes trials cluster  `Bin 1` `Bin 2`
#>        <int>  <dbl> <chr>      <dbl>   <dbl>
#>  1        35    100 Bin 2   8.89e-12   1.000
#>  2        37    100 Bin 2   1.13e-10   1.000
#>  3        36    100 Bin 2   3.17e-11   1.000
#>  4        38    100 Bin 2   4.03e-10   1.000
#>  5        46    100 Bin 2   1.05e- 5   1.000
#>  6        47    100 Bin 2   3.75e- 5   1.000
#>  7        40    100 Bin 2   5.12e- 9   1.000
#>  8        40    100 Bin 2   5.12e- 9   1.000
#>  9        44    100 Bin 2   8.28e- 7   1.000
#> 10        37    100 Bin 2   1.13e-10   1.000
#> # … with 90 more rows