Extract clustering assignments.

This function extracts the clustering assignments and the latent variables from the fit. This function must receive in input exactly the same data used for the fit.

Usage

Clusters(x, data)

Arguments

x

An object of class bmix that represents a fit.

data

The data used to compute the fit x.

Value

The input data transformed into a tibble, augmented with a column for the cluster labels and the latent variables.

Examples

# The same dataset used in the package vignette
data = data.frame(successes = c(rbinom(30, 100, .4), rbinom(70, 100, .7)), trials = 100)

# BMix fit with default parameters
x = bmixfit(data)
#> 
#> ── BMix fit ────────────────────────────────────────────────────────────────────
#> 
#> ℹ Binomials k_B = 1 and 2, Beta-Binomials k_BB = 0; 4 fits to run.
#> 
#> ℹ Bmix best fit completed in 0 mins
#> 
#> ── [ BMix ] My BMix model n = 100 with k = 2 component(s) (2 + 0) ──────────────
#> • Clusters: π = 70% [Bin 2] and 30% [Bin 1], with π > 0.
#> • Binomial Bin 1 with mean = 0.389612775724103.
#> • Binomial Bin 2 with mean = 0.69951992720296.
#> ℹ Score (model selection): ICL = 756.82.

Clusters(x, data)
#> # A tibble: 100 × 5
#>    successes trials cluster `Bin 1`  `Bin 2`
#>        <int>  <dbl> <chr>     <dbl>    <dbl>
#>  1        38    100 Bin 1      1.00 1.77e-10
#>  2        34    100 Bin 1      1.00 6.30e-13
#>  3        38    100 Bin 1      1.00 1.77e-10
#>  4        41    100 Bin 1      1.00 1.22e- 8
#>  5        40    100 Bin 1      1.00 2.97e- 9
#>  6        37    100 Bin 1      1.00 4.33e-11
#>  7        36    100 Bin 1      1.00 1.06e-11
#>  8        41    100 Bin 1      1.00 1.22e- 8
#>  9        32    100 Bin 1      1.00 3.76e-14
#> 10        38    100 Bin 1      1.00 1.77e-10
#> # … with 90 more rows